open/close now uses effects
